pyproject.toml
[project]
name = "mopenai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "openai>=1.64.0",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "typer>=0.15.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "pytest-asyncio>=0.25.3",
    "pytest-env>=1.1.5",
    "respx>=0.22.0",
    "ruff>=0.9.7",
]

[tool.ruff]
line-length = 120

exclude = ["excluded_file.py"]
lint.select = [
    "E",  # pycodestyle errors (settings from FastAPI, thanks, @tiangolo!)
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
lint.ignore = [
    "E501",  # line too long
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "preserve"

[tool.ruff.lint.isort]
order-by-type = true
relative-imports-order = "closest-to-furthest"
extra-standard-library = ["typing"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.pytest_env]
PYTEST_RUNNING = 1
PYTEST_ENV_FILE = ".env"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"